refactor: break down config parser.rs into focused modules
- Split 1753-line parser.rs into organized module hierarchy
- All production files under 200 lines (largest: 177 lines)
- Group by function: ffi/, types/, processing, validation
- Use directory hierarchies, no underscores
- Maintain all functionality and test coverage
