+++ crates/config/src/parser/types/commands.rs
@@ -1,0 +1,8 @@
+//! Command configuration types
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CommandConfig {
+    pub capabilities: Option<Vec<String>>,
+}
